Moved PhotonPolarization to utils BeyondDFTWorkflosParser
