+++ electronicparsers/utils/utils.py
@@ -26,7 +26,8 @@ from nomad.datamodel.metainfo.workflow import Workflow, Task, GW as GWold
 from nomad.datamodel.metainfo.workflow2 import Link, TaskReference
 from nomad.datamodel.metainfo.simulation.workflow import (
     SinglePoint, GW, GWMethod, GWResults, ParticleHoleExcitations,
-    ParticleHoleExcitationsMethod, ParticleHoleExcitationsResults
+    ParticleHoleExcitationsMethod, ParticleHoleExcitationsResults,
+    PhotonPolarization, PhotonPolarizationResults
 )
 
 
@@ -143,6 +144,46 @@ class BeyondDFTWorkflowsParser:
             workflow.tasks.append(task)
 
         gw_workflow_archive.workflow2 = workflow
+
+    def parse_photon_workflow(self):
+        workflow = PhotonPolarization(results=PhotonPolarizationResults())
+        workflow.name = 'PhotonPolarization'
+
+        input_structure = self.archive.run[-1].system[-1]
+        input_method = self.archive.run[-1].method[-1]
+        workflow.inputs = [
+            Link(name='Input structure', section=input_structure),
+            # TODO this should be put under PhotonPolarizationMethod e.g. as method_ref
+            Link(name='Input BSE methodology', section=input_method)]
+        spectra = []
+        outputs = []
+        for path in self._child_archives.keys():
+            archive = self._child_archives.get(path)
+            index = list(self._child_archives.keys()).index(path)
+            archive.workflow2 = SinglePoint()
+            archive.workflow2.name = 'Single point'
+
+            task = TaskReference(task=archive.workflow2)
+            input_photon_method = archive.run[-1].method[0]
+            if input_structure and input_photon_method:
+                archive.workflow2.inputs = [
+                    Link(name='Input structure', section=input_structure),
+                    Link(name='Input photon parameters', section=input_photon_method)]
+            output_calculation = archive.run[-1].calculation[-1]  # ref to EPSILON calculation
+            if output_calculation:
+                archive.workflow2.outputs = [Link(name=f'Output polarization {index + 1}', section=output_calculation)]
+                spectra.append(output_calculation.spectra[0])
+                outputs.append(Link(name=f'Output polarization {index + 1}', section=output_calculation))
+            archive.workflow2.tasks = [Task(name='Spectra calculation', inputs=archive.workflow2.inputs, outputs=archive.workflow2.outputs)]
+            workflow.tasks.append(task)
+            archive.metadata.entry_name = 'exciting Spectra Single Point workflow'
+
+        workflow.results.n_polarizations = len(spectra)
+        workflow.results.spectrum_polarization = spectra
+        outputs.append(Link(name='Workflow results', section=workflow.results))
+        workflow.outputs = outputs
+        self.archive.metadata.entry_name = 'exciting Photon Polarization workflow'
+        self.archive.workflow2 = workflow
 
     def parse_xs_workflow(self, xs_archives, xs_workflow_archive):
         sec_run = xs_workflow_archive.m_create(Run)
